move JobError exception classes to own file
