+++ api/scpca_portal/exceptions/__init__.py
@@ -3,3 +3,13 @@ from scpca_portal.exceptions.dataset_error import (
     DatasetLockedProjectError,
     DatasetMissingLibrariesError,
 )
+from scpca_portal.exceptions.job_error import (
+    JobError,
+    JobInvalidRetryStateError,
+    JobInvalidTerminateStateError,
+    JobSubmissionFailedError,
+    JobSubmitNotPendingError,
+    JobSyncNotProcessingError,
+    JobSyncStateFailedError,
+    JobTerminationFailedError,
+)
